requirements.txt
telethon==1.42.0
customtkinter==5.2.2
darkdetect==0.8.0
python-dotenv==1.2.1
packaging==26.0
requests==2.32.3
pyinstaller==6.19.0
